Refactor path generation for ctags_target_for_gcc_minus_e.cpp
Previously, the relative filename passed to GCCPreprocRunner() was made
absolute by prepareGCCPreprocRecipeProperties() and then returned, so it
could be used later on. There was an exception to this when /dev/null
was passed.

Now, the only place that passed something other than /dev/null simply
does this processing up front instead. This prepares the way for
removing Context::FileToRead in the next commit.

Signed-off-by: Matthijs Kooijman <matthijs@stdin.nl>